adding changes

+++ aks-bikeride-com/bikeride/src/main/java/com/bikeride/Service/impl/RideServiceImpl.java
@@ -5,6 +5,7 @@ import com.bikeride.Model.Request.JoinRideRequest;
 import com.bikeride.Model.Request.RideRequest;
 import com.bikeride.Model.Response.BikeRideResponse;
 import com.bikeride.Model.RideEvent;
+import com.bikeride.Model.User;
 import com.bikeride.Repository.RideRepository;
 import com.bikeride.Service.RideService;
 import lombok.extern.slf4j.Slf4j;
@@ -83,6 +84,10 @@ public class RideServiceImpl implements RideService {
     @Override
     public BikeRideResponse addUser(AddUserToRide addUserToRide) {
         BikeRideResponse bikeRideResponse=new BikeRideResponse();
+        bikeRideResponse=checkForAdminUser(addUserToRide.getRideId(),addUserToRide.getUserId(),"ADD");
+        if("2000".equalsIgnoreCase(bikeRideResponse.getResCode())){
+            return bikeRideResponse;
+        }
         int addUs=rideRepository.addUser(addUserToRide);
         if(addUs>0){
             bikeRideResponse.setResCode("2000");
@@ -94,9 +99,34 @@ public class RideServiceImpl implements RideService {
         return bikeRideResponse;
     }
 
+    private BikeRideResponse checkForAdminUser(String rideId, String userId, String command){
+        log.info("RideServiceImpl checkForAdminUser rideId:{}, userId:{}",rideId,userId);
+        BikeRideResponse bikeRideResponse=new BikeRideResponse();
+        bikeRideResponse.setResCode("2000");
+        bikeRideResponse.setResStatus("Success");
+        RideRequest rideRequest=new RideRequest();
+        rideRequest.setRideId(rideId);
+        RideEvent rideEvent=rideRepository.rideDetails(rideRequest);
+        if(rideEvent==null){
+            bikeRideResponse.setResCode("4001");
+            bikeRideResponse.setErrorDesc("Ride Not Found");
+        }else if(rideEvent.getUserId()!=userId){
+            bikeRideResponse.setResCode("4002");
+            bikeRideResponse.setErrorDesc("Riders created the ride can only add or remove people!");
+        }else if(rideEvent.getUser()!=null && rideEvent.getUser().size()==1 && "DELETE".equalsIgnoreCase(command)){
+            bikeRideResponse.setResCode("4003");
+            bikeRideResponse.setErrorDesc("Please delete the ride");
+        }
+        return bikeRideResponse;
+    }
+
     @Override
     public BikeRideResponse removeUser(AddUserToRide addUserToRide) {
         BikeRideResponse bikeRideResponse=new BikeRideResponse();
+        bikeRideResponse=checkForAdminUser(addUserToRide.getRideId(),addUserToRide.getUserId(),"DELETE");
+        if("2000".equalsIgnoreCase(bikeRideResponse.getResCode())){
+            return bikeRideResponse;
+        }
         int addUs=rideRepository.removeUser(addUserToRide);
         if(addUs>0){
             bikeRideResponse.setResCode("2000");
